risk-mitigation/risk_model.py
import numpy as np

class RiskModel:
    def __init__(self, data):
        self.data = data

    def calculate_returns(self):
        # Calculate returns based on the provided data
        returns = np.diff(self.data) / self.data[:-1]
        return returns

    def calculate_volatility(self):
        # Calculate volatility based on the returns
        returns = self.calculate_returns()
        volatility = np.std(returns)
        return volatility

    def analyze_risk(self):
        # Perform risk analysis using historical data
        returns = self.calculate_returns()
        volatility = self.calculate_volatility()

        # Add your risk analysis logic here
        # For example, you could calculate Value-at-Risk (VaR),
        # Conditional Value-at-Risk (CVaR), or other risk measures

        # Return the analysis results
        return {
            "returns": returns,
            "volatility": volatility,
            # Add more analysis results as needed
        }

    def mitigate_risk(self, risk_analysis):
        # Develop a risk mitigation plan based on the analysis results
        # For example, you could adjust asset allocations,
        # implement hedging strategies, or rebalance portfolios

        # Return the mitigation plan
        return {
            # Define the mitigation plan based on the analysis results
        }
