tests/test_risk_model.py
import numpy as np
import unittest
from risk_mitigation.risk_model import RiskModel

class TestRiskModel(unittest.TestCase):
    def setUp(self):
        self.data = np.array([100, 110, 105, 115, 120])

    def test_analyze_risk(self):
        risk_model = RiskModel(self.data)
        result = risk_model.analyze_risk()

        self.assertIsInstance(result, dict)
        self.assertIn("returns", result)
        self.assertIn("volatility", result)

    def test_mitigate_risk(self):
        risk_model = RiskModel(self.data)
        risk_analysis = risk_model.analyze_risk()
        mitigation_plan = risk_model.mitigate_risk(risk_analysis)

        self.assertIsInstance(mitigation_plan, dict)
        # Add more assertions to verify the mitigation plan

if __name__ == '__main__':
    unittest.main()
